chore: move more stuff
