+++ crates/storage/db-api/Cargo.toml
@@ -14,9 +14,11 @@ workspace = true
 [dependencies]
 # reth
 reth-codecs.workspace = true
+reth-fs-util.workspace = true
 reth-primitives.workspace = true
 reth-prune-types.workspace = true
 reth-storage-errors.workspace = true
+reth-tracing.workspace = true
 
 # codecs
 modular-bitfield.workspace = true
@@ -29,6 +31,7 @@ metrics.workspace = true
 # misc
 derive_more.workspace = true
 bytes.workspace = true
+sysinfo = "0.30"
 
 # arbitrary utils
 arbitrary = { workspace = true, features = ["derive"], optional = true }
@@ -58,6 +61,7 @@ proptest.workspace = true
 proptest-derive.workspace = true
 
 paste.workspace = true
+tempfile.workspace = true
 
 assert_matches.workspace = true
