+++ crates/storage/db-api/src/lib.rs
@@ -69,6 +69,7 @@ pub mod cursor;
 pub mod database;
 /// Database metrics trait extensions.
 pub mod database_metrics;
+pub mod lockfile;
 pub mod mock;
 /// Table traits
 pub mod table;
