+++ crates/storage/db-mdbx/src/lib.rs
@@ -15,6 +15,7 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+mod metrics;
 mod utils;
 
 pub mod mdbx;
@@ -22,6 +23,7 @@ pub mod mdbx;
 pub use reth_storage_errors::db::{DatabaseError, DatabaseWriteOperation};
 
 pub use mdbx::{create_db, init_db, open_db, open_db_read_only, DatabaseEnv, DatabaseEnvKind};
+pub use utils::is_database_empty;
 
 pub use reth_db_api::*;
