+++ crates/storage/db/src/lib.rs
@@ -15,8 +15,6 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-pub mod lockfile;
-mod metrics;
 pub mod static_file;
 pub mod tables;
 mod utils;
